+++ riscv/insn_template.h
@@ -1,5 +1,6 @@
 // See LICENSE for license details.
 
+#include "decode_macros.h"
 #include "arith.h"
 #include "mmu.h"
 #include "softfloat.h"
