+++ riscv/rom.cc
@@ -1,3 +1,4 @@
+#include "common.h"
 #include "devices.h"
 
 rom_device_t::rom_device_t(std::vector<char> data)
